Add support for default branches
How it works:
Repos without a branch specified will get the *magic branch* ___. When
updating a repo with this branch the head of origin will be updated.
When this fails (the head branch isn't known locally) the error message
is parsed for the new heads name. Then the new head is fetched and
origin updated again.
Afterwards the symbolic ref for origins head is resolved and checked if
it is the current branch.
If not the new branch is deleted (in case it existed), the new head is
checked out and the old head branch is deleted.

Closes #16