Move bins from sources/_/bin to bin
After merging #18 it doesn't make too much sense add bins to the sources
folder. So the bins will now be stored in the zgenom project path.
This was the initial idea as well but was a bit complicated since a
github username could be `bin` as well.

When updating it may result in residual data in `$ZGEN_DIR/_/bin`. Since
there are only links stored it should be negligible. Please remove this
folder manually if it bothers you.