Add fpath before sourcing the plugin
If the plugin wants to autoload a function it has to add itself to
`fpath`. Ideally the plugin manager should take care of this.

Now zgenom adds modifies the fpath before sourcing the plugin. To keep
it consistent the order of fpaths is now inverted since now the first
`load` is added the first instead of prepended at the first position.

Closes https://github.com/tarjoilija/zgen/issues/105