Improve folder structure
Before a folder was named `user/repo-branch`.
Given the user `jane-doe`, repo `my-plugin` and branch `the-branch` this
folder would be created:

`jane-doe/my-plugin-the-branch`

There is no way to parse the branch from the folder name.
With the new format it is named `jane-doe/my-plugin/the-branch` making
it easy to parse since slashes are escaped.

The next time after `zgenom reset` the repos will be moved
automatically.