Don't use env vars to control autoupdates
Since the autoupdate commands are going to need to be run explicitly as
`zgen autoupdate-{system,plugins}`, we're going to have them called with
a single integer argument, the number of days allowed to elapse between
updates.

Break up the autoupdate functionality into separate commands for
updating the system and the plugins to give users more control.