MAYBE BREAKING: Load `init.zsh` on `zgenom saved`
Instead of always loading the `init.zsh` when `zgenom.zsh` is sourced,
now the init file is only loaded when `zgenom save` is called.

- If you use zgenom as described in the README you don't need to change
  anything.
- If you don't call `zgenom saved` you need to call `zgenom init` after
  you sourced `zgenom.zsh`.

This change is mainly done to make the new `zgenom autoupdate` working
correctly.
But maybe this will also unlock some new possible features since it is
now possible for the user to call functions of zgenom before any plugin
is loaded. (Which was previously only possible when no `init.zsh` is
present.)