Refactor zgenom to a new architecture
Instead of defining all functions in one large file every main api is
defined in its own file.
This should lead to easier maintenance and only the needed parts are
loaded into the current shell - making it potentially faster.
But the main focus is to find the code faster without searching.