Use a default path for zcompdump
On some machines the default compdump name contains the machines name.
On Macs this includes `’` by default which results in errors when
compiling the file. Apart from the error message it also slows down the
startup time since the uncompiled compdump file is way slower to load.

This closes #100, #109 and #121