Use a default path for zcompdump (#122)
On some machines the default compdump name contains the machines name.
On Macs this includes `’` by default which results in errors when
compiling the file. But there are other invalid characters as well (e.g. `–`).
Apart from the error message it also slows down the startup time since the
uncompiled compdump file is way slower to load.

This closes #100, #109 and #121